VideoActivity to VideoFragment
