+++ app/src/main/java/com/losextraditables/bu/BuApplication.java
@@ -9,6 +9,7 @@ import com.losextraditables.bu.login.LoginModule;
 import com.losextraditables.bu.main.MainModule;
 import com.losextraditables.bu.pictures.PicturesModule;
 import com.losextraditables.bu.utils.AuthenticationHandler;
+import com.losextraditables.bu.videos.VideosModule;
 import dagger.ObjectGraph;
 import io.fabric.sdk.android.Fabric;
 import java.util.Arrays;
@@ -30,7 +31,7 @@ public class BuApplication extends RosieApplication {
 
   @Override protected List<Object> getApplicationModules() {
     return Arrays.asList(new ApplicationModule(this), new InstagrammersListModule(),
-        new LoginModule(), new PicturesModule(), new MainModule());
+        new LoginModule(), new PicturesModule(), new MainModule(), new VideosModule());
   }
 
   public void replaceGraph(ObjectGraph objectGraph) {
