+++ app/src/main/java/com/losextraditables/bu/main/MainModule.java
@@ -1,11 +1,12 @@
 package com.losextraditables.bu.main;
 
 import com.losextraditables.bu.pictures.view.activity.PicturesFragment;
+import com.losextraditables.bu.videos.view.activity.VideoFragment;
 import dagger.Module;
 
 @Module(library = true,
     complete = false,
     injects = {
-        MainTabbedActivity.class, PicturesFragment.class
+        MainTabbedActivity.class, PicturesFragment.class, VideoFragment.class
     }) public class MainModule {
 }
