+++ app/src/main/res/layout/content_video.xml
@@ -6,7 +6,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     app:layout_behavior="@string/appbar_scrolling_view_behavior"
-    tools:context="com.losextraditables.bu.videos.view.activity.VideoActivity"
+    tools:context=".videos.view.activity.VideoFragment"
     tools:showIn="@layout/activity_video"
     >
